[한성태] sprint4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -1,11 +1,12 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.FindUserDto;
-import com.sprint.mission.discodeit.dto.SaveUserParamDto;
-import com.sprint.mission.discodeit.dto.UpdateUserParamDto;
+import com.sprint.mission.discodeit.dto.*;
+import com.sprint.mission.discodeit.entity.BinaryContent;
+import com.sprint.mission.discodeit.entity.BinaryData;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
+import com.sprint.mission.discodeit.repository.BinaryDataRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserService;
@@ -13,6 +14,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -23,46 +26,50 @@ public class BasicUserService implements UserService {
     private final UserRepository userRepository;
     private final UserStatusRepository userStatusRepository;
     private final BinaryContentRepository binaryContentRepository;
+    private final BinaryDataRepository binaryDataRepository;
 
     @Override
-    public void save(SaveUserParamDto saveUserParamDto) {
+    public void save(SaveUserRequestDto saveUserParamDto, Optional<SaveBinaryContentRequestDto> saveBinaryContentRequestDto) {
         if (userRepository.findUserByUsername(saveUserParamDto.username()).isPresent()) {
-            System.out.println("[실패] 회원아이디 중복");
-            return;
+            throw new IllegalArgumentException("사용자 이름 중복");
         }
 
         if (userRepository.findUserByEmail(saveUserParamDto.email()).isPresent()) {
-            System.out.println("[실패] 회원이메일 중복");
-            return;
+            throw new IllegalArgumentException("이메일 중복");
         }
+
+        UUID profileId = saveBinaryContentRequestDto
+                .map(profile -> {
+                    String fileName = profile.fileName();
+                    String contentType = profile.contentType();
+                    byte[] fileData = profile.fileData();
+                    BinaryData binaryData = binaryDataRepository.save(new BinaryData(fileData));
+                    BinaryContent binaryContent = new BinaryContent(binaryData.getId(), fileName, contentType);
+                    binaryContentRepository.save(binaryContent);
+                    return binaryContent.getId();
+                })
+                .orElse(null);
 
         User user = new User(
                 saveUserParamDto.username(), saveUserParamDto.password(),
                 saveUserParamDto.nickname(), saveUserParamDto.email(),
-                saveUserParamDto.profileUUID());
+                profileId);
         userRepository.save(user);
         UserStatus userStatus = UserStatus.builder()
                 .userUUID(user.getId())
                 .build();
         userStatusRepository.save(userStatus);
-
-        if (user == null) {
-            System.out.println("[실패] 저장 실패.");
-            return;
-        }
-
-        System.out.println("[성공] 회원가입 성공");
     }
 
     @Override
-    public FindUserDto findByUser(UUID userUUID) {
-        User user = userRepository.findUserById(userUUID)
+    public FindUserDto findByUser(UUID userId) {
+        User user = userRepository.findUserById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원"));
         UserStatus userStatus = userStatusRepository.findByUserId(user.getId())
                 .orElseThrow(() -> new IllegalArgumentException("사용자 상태 확인 불가"));
 
         FindUserDto findUserDto = new FindUserDto(
-                user.getId(), user.getNickname(),
+                user.getId(), user.getUsername(), user.getEmail(),
                 user.getProfile(), user.getCreatedAt(),
                 user.getUpdatedAt(), userStatus.getUpdatedAt(),
                 userStatus.isLastStatus());
@@ -80,25 +87,37 @@ public class BasicUserService implements UserService {
     }
 
     @Override
-    public void update(UpdateUserParamDto updateUserDto) {
-        User user = userRepository.findUserById(updateUserDto.userUUID()).
+    public void update(UUID userId, UpdateUserRequestDto updateUserDto, Optional<SaveBinaryContentRequestDto> saveBinaryContentRequestDto) {
+        User user = userRepository.findUserById(userId).
                 orElseThrow(() -> new NullPointerException("사용자가 존재하지 않습니다."));
-        userRepository.update(updateUserDto.userUUID(), updateUserDto.nickname(), updateUserDto.profileId());
+
+        user.updateNickname(updateUserDto.nickname());
+        UUID profileId = saveBinaryContentRequestDto
+                .map(profile -> {
+                    String fileName = profile.fileName();
+                    String contentType = profile.contentType();
+                    byte[] fileData = profile.fileData();
+                    BinaryData binaryData = binaryDataRepository.save(new BinaryData(fileData));
+                    BinaryContent binaryContent = new BinaryContent(binaryData.getId(), fileName, contentType);
+                    binaryContentRepository.save(binaryContent);
+                    return binaryContent.getId();
+                })
+                .orElse(null);
+        user.updateProfile(profileId);
+        userRepository.save(user);
     }
 
     @Override
-    public void delete(UUID userUUID) {
-        User user = userRepository.findUserById(userUUID)
-                .orElseThrow(NullPointerException::new);
-        binaryContentRepository.delete(user.getProfile());
+    public void delete(UUID userId) {
+        User user = userRepository.findUserById(userId)
+                .orElseThrow(() -> new NullPointerException("사용자가 존재하지 않습니다."));
+        if(Objects.nonNull(user.getProfile())) {
+            binaryContentRepository.delete(user.getProfile());
+        }
         UserStatus userStatus = userStatusRepository.findByUserId(user.getId())
                 .orElseThrow(() -> new IllegalArgumentException("사용자 상태 오류"));
+
         userStatusRepository.delete(userStatus.getId());
-        boolean isDeleted = userRepository.deleteUserById(user.getId());
-        if (!isDeleted) {
-            System.out.println("삭제 실패");
-            return;
-        }
-        System.out.println("[성공]");
+        userRepository.delete(user.getId());
     }
 }
